+++ examples/ReinforcementLearning/ReinforcementLearning/GameConfigurations/SpaceInvadersConfiguration.cs
