+++ examples/ReinforcementLearning/ReinforcementLearning/ReplayMemory.cs
@@ -6,7 +6,7 @@ using Newtonsoft.Json;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
 
-namespace AtariDeepQLearner
+namespace ReinforcementLearning
 {
     public class Episode
     {
